refactor: move components to specific directories
