+++ packages/rockets-next-web/src/app/(auth)/layout.tsx
@@ -4,7 +4,7 @@ import { ReactNode } from "react";
 import useCheckTokens from "@/hooks/useCheckTokens";
 import { useRouter } from "next/navigation";
 
-import Loading from "@/components/Loading/Loading";
+import Loading from "@/components/shared/Loading";
 
 export default function LayoutAuth({ children }: { children: ReactNode }) {
   const router = useRouter();
