+++ packages/rockets-next-web/src/app/(auth)/reset-password/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ResetPasswordScreen from "@/components/ResetPasswordScreen/ResetPasswordScreen";
+import ResetPasswordScreen from "@/components/auth/ResetPasswordScreen";
 
 const ResetPassword = () => {
   return <ResetPasswordScreen />;
