+++ packages/rockets-next-web/src/app/(auth)/sign-up/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import SignScreen from "@/components/SignScreen/SignScreen";
+import SignScreen from "@/components/auth/SignScreen";
 
 const SignUp = () => {
   return <SignScreen isSignUp />;
